Introduce /continuous, various tweaks
First I wanted to create something more of a "desktop" base, distinct
from all of the "Workstation" stuff.  Particularly, a goal with this
should be to *not* install `make` and `gcc` on the host system, but
only in containers.

So the workstation ref inherits from desktop for now.

Also in order to actively produce something useful, we really need to
skip the whole koji/bodhi crap for things that we need to actively
hack on (at least for now) and jump to continuous delivery from git.

I used as a basis the work in
https://wiki.centos.org/SpecialInterestGroup/Atomic/Devel to add
automated builds of this into CentOS CI, currently under the Atomic
SIG.

Regarding content: a major thing I had to battle is that some things
depend on yum/dnf, so I ended up creating:
https://github.com/cgwalters/yum-stub

I also had to drop gnome-software since it depends on PackageKit
which conflicts with the version of libhif.

For this we should do a different build of software without PK.

Other changes:
 - Added docker